requirements.txt
flask
transformers
ollama
pdfplumber
python-docx
